requirements.txt
PyQt5
openai-whisper
python-docx
ffmpeg-python
